+++ server/core/map.js
@@ -1,15 +1,18 @@
+const PF = require('pathfinding');
+const config = require('./../core/config');
 const surfaceMap = require('./../maps/layers/surface.json');
 const world = require('./../core/world');
+const UI = require('./../core/utilities/ui');
+const { droppedItems } = require('./../data/default');
 
 class Map {
   constructor(level) {
     // Getters & Setters
+    this.players = [];
     this.level = level;
 
     this.background = world.map.background;
     this.foreground = world.map.foreground;
-
-    this.players = [];
 
     this.setUp();
   }
@@ -71,11 +74,19 @@ class Map {
     }
   }
 
+  /**
+   * Set up the map
+   */
   async setUp() {
+    // Load the board
     const board = await Map.load();
 
+    // Set background and foreground tile data
     this.background = board[0].data;
     this.foreground = board[1].data;
+
+    // Set items on map
+    world.items = droppedItems;
   }
 
   /**
@@ -91,6 +102,53 @@ class Map {
 
     return new Promise((resolve) => {
       resolve(mapToLoad[level].layers);
+    });
+  }
+
+  /**
+   * Get the blocked/non-blocked tile-matrix of their viewport
+   *
+   * @param {object} player The player asking
+   */
+  static getMatrix(player) {
+    const x = player.x;
+    const y = player.y;
+
+    const { size, viewport } = config.map;
+
+    return new Promise((resolve) => {
+      const tileCrop = {
+        x: x - Math.floor(0.5 * viewport.x),
+        y: y - Math.floor(0.5 * viewport.y),
+      };
+
+      const matrix = [];
+
+      // Drawing the map row by column.
+      for (let column = 0; column <= viewport.y; column += 1) {
+        const grid = [];
+        for (let row = 0; row <= viewport.x; row += 1) {
+          const tileToFind = (((column + tileCrop.y) * size.x) + row) + tileCrop.x;
+          const backgroundTile = world.map.background[tileToFind] - 1;
+          const foregroundTile = (world.map.foreground[tileToFind] - 1) - 252;
+          // 252 because of the gid problem in Tiled
+
+          // Is the background walkable?
+          let walkable = UI.tileWalkable(backgroundTile) ? 0 : 1;
+          // If it is not, is the foreground walkable?
+          if (walkable === 0) walkable = UI.tileWalkable(foregroundTile, 'foreground') ? 0 : 1;
+
+          // Push the block/non-blocked tile to the
+          // grid so that the pathfinder can use it
+          grid.push(walkable);
+        }
+
+        // Push blocked/non-blocked array for pathfinding
+        matrix.push(grid);
+      }
+
+      // The new walkable/non-walkable grid
+      resolve(new PF.Grid(matrix));
     });
   }
 }
