+++ server/core/npc.js
@@ -1,3 +1,11 @@
+const emoji = require('node-emoji');
+
+const Socket = require('./../socket');
+const UI = require('./utilities/ui');
+const world = require('./world');
+const npcs = require('./../data/npcs');
+
+
 class NPC {
   constructor(data) {
     this.id = data.id;
@@ -27,6 +35,86 @@ class NPC {
 
     // What column they are on in tileset
     this.column = data.graphic.column;
+  }
+
+  /**
+   * Load the NPCs into the game world
+   *
+   * @param {this} context The server context
+   */
+  static load(context) {
+    npcs.forEach((npc) => {
+      world.npcs.push(new NPC(npc));
+    }, context);
+    console.log(`${emoji.get('walking')}  Loading NPCs...`);
+  }
+
+  /**
+   * Simulate NPC movement every cycle
+   */
+  static movement() {
+    world.npcs.map((npc) => {
+      // Next movement allowed in 2.5 seconds
+      const nextActionAllowed = npc.lastAction + 2500;
+
+      if (npc.lastAction === 0 || nextActionAllowed < Date.now()) {
+        // Are they going to move or sit still this time?
+        const action = UI.getRandomInt(1, 2) === 1 ? 'move' : 'nothing';
+
+        // NPCs going to move during this loop?
+        if (action === 'move') {
+          // Which way?
+          const direction = ['up', 'down', 'left', 'right'];
+          const going = direction[UI.getRandomInt(0, 3)];
+
+          // What tile will they be stepping on?
+          const tile = {
+            background: UI.getFutureTileID(world.map.background, npc.x, npc.y, going),
+            foreground: UI.getFutureTileID(world.map.foreground, npc.x, npc.y, going),
+          };
+
+          switch (going) {
+            default:
+            case 'up':
+              if ((npc.y - 1) >= (npc.spawn.y - npc.range)) {
+                if (UI.tileWalkable(tile.background) && UI.tileWalkable(tile.foreground)) {
+                  npc.y -= 1;
+                }
+              }
+              break;
+            case 'down':
+              if ((npc.y + 1) <= (npc.spawn.y + npc.range)) {
+                if (UI.tileWalkable(tile.background) && UI.tileWalkable(tile.foreground)) {
+                  npc.y += 1;
+                }
+              }
+              break;
+            case 'left':
+              if ((npc.x - 1) >= (npc.spawn.x - npc.range)) {
+                if (UI.tileWalkable(tile.background) && UI.tileWalkable(tile.foreground)) {
+                  npc.x -= 1;
+                }
+              }
+              break;
+            case 'right':
+              if ((npc.x + 1) <= (npc.spawn.x + npc.range)) {
+                if (UI.tileWalkable(tile.background) && UI.tileWalkable(tile.foreground)) {
+                  npc.x += 1;
+                }
+              }
+              break;
+          }
+        }
+
+        // Register their last action
+        npc.lastAction = Date.now();
+      }
+
+      return npc;
+    });
+
+    // Tell the clients of the new NPCs
+    Socket.broadcast('npc:movement', world.npcs);
   }
 }
